--- frontend/src/pages/PendingAssessments/RequestItem/RequestButtons/types/index.ts
@@ -1,13 +1,0 @@
-import { IManageAssessmentDto } from 'types/api.dto';
-
-export interface IAssessmentRequestButtonsContainerProps {
-  id: string;
-  manageAssessment: (payload: IManageAssessmentDto) => void;
-  isActionLoading: boolean;
-}
-
-export interface IAssessmentRequestButtonsProps {
-  isActionLoading: boolean;
-  approveAssessment: () => void;
-  declineAssessment: () => void;
-}
