+++ frontend/src/pages/PendingAssessments/RequestItem/RequestTechnologies/RequestTechnologies.tsx
@@ -1,7 +1,8 @@
 import { FC } from 'react';
 
+import { IRequestTechnologiesProps } from 'pages/PendingAssessments/types';
+
 import { TechnologiesWrapper } from './styled';
-import { IRequestTechnologiesProps } from './types';
 import TechnologyItem from './TechnologyItem';
 
 const RequestTechnologies: FC<IRequestTechnologiesProps> = ({ technologies }) => (
