--- frontend/src/pages/PendingAssessments/RequestItem/RequestTechnologies/TechnologyItem/types/index.ts
@@ -1,12 +1,0 @@
-import { ITechnology } from 'types/skill';
-
-export interface ITechnologyItemProps {
-  technology: ITechnology;
-  isInfoShown: boolean;
-  showInfo: () => void;
-  hideInfo: () => void;
-}
-
-export interface ITechnologyItemContainerProps {
-  technology: ITechnology;
-}
