--- frontend/src/pages/PendingAssessments/RequestItem/RequestTechnologies/types/index.ts
@@ -1,5 +1,0 @@
-import { IAssessment } from 'types/request';
-
-export interface IRequestTechnologiesProps {
-  technologies: IAssessment['course']['technologies'];
-}
