--- frontend/src/pages/PendingAssessments/RequestItem/types/index.ts
@@ -1,11 +1,0 @@
-import { IManageAssessmentDto } from 'types/api.dto';
-import { IAssessment } from 'types/request';
-
-export interface IAssessmentRequestItemProps {
-  user: IAssessment['user'];
-  course: IAssessment['course'];
-  elapsed: IAssessment['elapsed'];
-  clientCourseId: IAssessment['_id'];
-  manageAssessment: (payload: IManageAssessmentDto) => void;
-  isActionLoading: boolean;
-}
