--- frontend/src/pages/PendingAssessments/types/index.ts
@@ -1,9 +1,0 @@
-import { IManageAssessmentDto } from 'types/api.dto';
-import { IAssessment } from 'types/request';
-
-export interface IPendingAssessmentsProps {
-  isLoading: boolean;
-  isActionLoading: boolean;
-  manageAssessment: (payload: IManageAssessmentDto) => void;
-  assessments?: IAssessment[];
-}
